refactor(ims): reuse parsers and consolidate SMS regression tests
Reuse the existing BER length decoder and port validation, replace manual address splitting with net.SplitHostPort, and remove unused route metadata and the obsolete random UUID helper. Keep stable instance and MO identity policies in their existing modules without changing their behavior.

Consolidate focused regressions into existing test suites, remove duplicate setup and subprocess scaffolding, and retain real loopback SMS/RPDU and cancellation coverage. Verified against the original pre-refactor regression suite and a local 100,000-case URI acceptance comparison.
