+++ internal/vowifi/ims/message_identity_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 	"time"
 
@@ -127,5 +128,165 @@ func TestDeliveryReportUsesAssociatedIdentityOnWire(t *testing.T) {
 				t.Fatal("MESSAGE selection changed the REGISTER identity")
 			}
 		})
+	}
+}
+
+func TestMOSMSIdentityProvenanceOnWire(t *testing.T) {
+	// The same URI spelling is REGISTER-only when generated, but remains an
+	// eligible originating identity when explicitly configured.
+	const current = "sip:001010123456789@ims.mnc001.mcc001.3gppnetwork.org"
+	const networkDefault = "sip:12025550100@msg.example.test"
+	for _, tc := range []struct {
+		name, configured, want string
+	}{
+		{"generated", "", networkDefault},
+		{"configured", current, current},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			s, seen, _ := smsPSITestSession(t, &smsPSITestAKA{recordingAKA: &recordingAKA{result: vowifi.AKAResult{RES: []byte{1, 2, 3, 4}}}, psi: smsPSITestURI}, tc.configured)
+			if s.identity.public != current {
+				t.Fatalf("fixture identity = %q, want %q", s.identity.public, current)
+			}
+			s.mu.Lock()
+			s.evidence.AssociatedIdentities = []string{"<" + networkDefault + ">", "<" + current + ">"}
+			s.mu.Unlock()
+			r := smsPSITestSubmit(t, s, seen)
+			if publicIdentityURI(r.value("From")) != tc.want || publicIdentityURI(r.value("P-Preferred-Identity")) != tc.want {
+				t.Fatalf("wrong origin: From=%q PPI=%q", r.value("From"), r.value("P-Preferred-Identity"))
+			}
+			if r.URI != smsPSITestURI {
+				t.Fatal("PSI target changed")
+			}
+		})
+	}
+}
+
+func TestMOIdentityChangeDoesNotAffectRepliesOrUSSD(t *testing.T) {
+	for _, test := range []struct {
+		name, reply, content, preferred string
+	}{
+		{"reply", "network-call", smsContentType, ""},
+		{"explicit_identity", "", smsContentType, "current"},
+		{"ussd", "", ussiContentType, ""},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			// An eligible explicitly configured identity is not the generated
+			// REGISTER-only identity. Replies/USSD still exercise generated sessions.
+			configured := ""
+			if test.preferred != "" {
+				configured = "sip:subscriber@msg.example.test"
+			}
+			s, seen, _ := smsPSITestSession(t, &recordingAKA{result: vowifi.AKAResult{RES: []byte{1, 2, 3, 4}}}, configured)
+			current := s.identity.public
+			s.mu.Lock()
+			s.evidence.AssociatedIdentities = []string{"<sip:12025550100@msg.example.test>", "<" + current + ">"}
+			s.mu.Unlock()
+			preferred := test.preferred
+			if preferred == "current" {
+				preferred = current
+			}
+			_, e := s.sendSIPMessageWithIdentity(context.Background(), smsPSITestURI, []byte{3, 0}, test.reply, test.content, "smsip", preferred)
+			if e != nil {
+				t.Fatal(e)
+			}
+			select {
+			case r := <-seen:
+				if !strings.Contains(r.value("From"), current) || !strings.Contains(r.value("P-Preferred-Identity"), current) {
+					t.Fatal("non-MO identity changed")
+				}
+			case <-time.After(time.Second):
+				t.Fatal("no message")
+			}
+		})
+	}
+}
+
+func TestMOTemporaryPreferredIdentityCannotBypassPolicy(t *testing.T) {
+	for _, tc := range []struct {
+		name          string
+		usableDefault bool
+	}{
+		{"no_default", false},
+		{"usable_default", true},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			s, seen, _ := smsPSITestSession(t, &recordingAKA{result: vowifi.AKAResult{RES: []byte{1, 2, 3, 4}}})
+			current := s.identity.public
+			const valid = "sip:12025550100@msg.example.test"
+			s.mu.Lock()
+			s.evidence.AssociatedIdentities = []string{"<" + current + ">"}
+			if tc.usableDefault {
+				s.evidence.AssociatedIdentities = []string{"<" + valid + ">", "<" + current + ">"}
+			}
+			s.mu.Unlock()
+			_, err := s.sendSIPMessageWithIdentity(context.Background(), smsPSITestURI, []byte{0, 0}, "", smsContentType, "smsip", current)
+			if !tc.usableDefault {
+				if err == nil {
+					t.Error("explicit preferred temporary identity bypassed error")
+				}
+				select {
+				case <-seen:
+					t.Error("sent forbidden MO MESSAGE")
+				case <-time.After(30 * time.Millisecond):
+				}
+				return
+			}
+			if err != nil {
+				t.Fatal(err)
+			}
+			select {
+			case r := <-seen:
+				if publicIdentityURI(r.value("From")) != valid || publicIdentityURI(r.value("P-Preferred-Identity")) != valid {
+					t.Errorf("preferred temporary leaked: %v", r.Headers)
+				}
+			case <-time.After(time.Second):
+				t.Fatal("no MESSAGE")
+			}
+		})
+	}
+}
+
+func TestOriginatingSMSPublicIdentityRejectsUnusableDefault(t *testing.T) {
+	const temporary = "sip:001010123456789@ims.mnc001.mcc001.3gppnetwork.org"
+	const valid = "sip:12025550100@msg.example.test"
+	for _, tc := range []struct {
+		name       string
+		associated []string
+	}{
+		{"absent", nil},
+		{"same_temporary", []string{"<" + temporary + ">"}},
+		{"temporary_uri_parameter", []string{"<" + temporary + ";user=phone>"}},
+		{"temporary_escaped_digit", []string{"<sip:%3001010123456789@ims.mnc001.mcc001.3gppnetwork.org>"}},
+		{"same_temporary_case", []string{"<SIP:001010123456789@IMS.MNC001.MCC001.3GPPNETWORK.ORG>"}},
+		{"temporary_first_not_second", []string{"<" + temporary + ">, <" + valid + ">"}},
+		{"invalid_first_not_second", []string{"garbage", "<" + valid + ">"}},
+		{"invalid_sip", []string{"<sip:a@@example.test>"}},
+		{"invalid_tel", []string{"<tel:not-a-number>"}},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			got, _, err := originatingSMSPublicIdentity(temporary, tc.associated)
+			if err == nil || got != "" || !strings.Contains(err.Error(), "public identity") {
+				t.Fatalf("identity = %q, %v; want rejection", got, err)
+			}
+		})
+	}
+}
+
+func TestMOTemporaryIdentityWithoutDefaultNeverSends(t *testing.T) {
+	s, seen, _ := smsPSITestSession(t, &recordingAKA{result: vowifi.AKAResult{RES: []byte{1, 2, 3, 4}}})
+	s.mu.Lock()
+	s.evidence.AssociatedIdentities = nil
+	s.mu.Unlock()
+	result, err := s.SendSMS(context.Background(), vowifi.SMSSubmitRequest{Recipient: "+12025550123", Text: "OFFLINE TEST"})
+	if err == nil || !strings.Contains(err.Error(), "public identity") {
+		t.Errorf("want identity error, got %v", err)
+	}
+	if result.PartsAttempted != 0 {
+		t.Errorf("attempted %d parts without usable default", result.PartsAttempted)
+	}
+	select {
+	case r := <-seen:
+		t.Errorf("sent forbidden MO MESSAGE: From=%q", r.value("From"))
+	case <-time.After(30 * time.Millisecond):
 	}
 }
